TASK: refactor createBattle
